+++ src/main/java/real/world/security/service/CustomUserDetailsService.java
@@ -1,4 +1,4 @@
-package real.world.security;
+package real.world.security.service;
 
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.userdetails.UserDetails;
