+++ src/main/java/real/world/security/support/NorRequestMatcher.java
@@ -1,4 +1,4 @@
-package real.world.security;
+package real.world.security.support;
 
 import static java.util.stream.Collectors.toList;
