+++ src/main/java/real/world/security/support/RootNameObjectMapper.java
@@ -1,4 +1,4 @@
-package real.world.security;
+package real.world.security.support;
 
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
